requirements.txt
ipykernel==6.25.2
ipython==8.15.0
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.12.2
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
